+++ src/main/java/org/joelson/cts/calculator/model/Amount.java
@@ -1,12 +1,12 @@
 package org.joelson.cts.calculator.model;
 
-public record Amount(String currency, float amount) {
+public record Amount(String currency, double amount) {
 
-    public Amount multiplyBy(float f) {
+    public Amount multiplyBy(double f) {
         return new Amount(currency, amount * f);
     }
 
-    public Amount divideBy(float f) {
+    public Amount divideBy(double f) {
         return new Amount(currency, amount / f);
     }
